Criada a estrutura inicial do ClienteMapper

+++ Project/include/model/Promocao.h
@@ -15,7 +15,7 @@ private:
 public:
     Promocao(int id, double percentagem , const std::string& data_inicio , const std::string& data_final, Produto* produto, Categoria* categoria);
     int getId() const;
-    int getPercentagem() const;
+    double getPercentagem() const;
     std::string getData_Inicio() const;
     std::string getData_Final() const;
     Produto* getProduto() const;
